Добавил контроллер для доски задач
